refactor: move events utils to a new file
