+++ src/plugin.js
@@ -1,12 +1,13 @@
 import dictionary from './dictionary';
 import mixin from './mixin';
 import directive from './directive';
-import { assign, getPath, warn, isCallable, detectPassiveSupport } from './utils';
+import { assign, getPath, warn, isCallable } from './utils';
 import Validator from './core/validator';
 import ErrorBag from './core/errorBag';
 import mapFields from './core/mapFields';
 import { ValidationProvider } from './components';
 import I18nDictionary from './core/localization/i18n';
+import { detectPassiveSupport } from './utils/events';
 
 // @flow
